Moved common logic of ISO handling to static helper. Implemented more "fast" methods. Infrastructure package updates.
