+++ src/NodaTime.Serialization.Utf8Json/DS.NodaTime.Serialization.Utf8Json.Tests/DS.NodaTime.Serialization.Utf8Json.Tests.csproj
@@ -6,7 +6,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="coverlet.msbuild" Version="2.4.0" />
+      <PackageReference Include="coverlet.msbuild" Version="2.5.0" />
       <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
       <PackageReference Include="NodaTime" Version="2.0.0" />
       <PackageReference Include="Utf8Json" Version="1.3.7" />
