+++ src/NodaTime.Serialization.Utf8Json/DS.NodaTime.Serialization.Utf8Json/NodaLocalDateFormatter.cs
@@ -1,6 +1,7 @@
 using System;
 using NodaTime;
 using Utf8Json;
+using static DS.NodaTime.Serialization.Utf8Json.Helpers.NodaIsoDateTimeHandler;
 
 namespace DS.NodaTime.Serialization.Utf8Json
 {
@@ -16,82 +17,22 @@ namespace DS.NodaTime.Serialization.Utf8Json
         public void Serialize(ref JsonWriter writer, LocalDate value, IJsonFormatterResolver formatterResolver)
         {
             _validator?.Invoke(value);
+            const int length = 10 + 2; // {YEAR}-{MONTH}-{DAY} + quotation
+
+            writer.EnsureCapacity(length);
+
             writer.WriteRawUnsafe((byte)'\"');
-            var year = value.Year;
-            var month = value.Month;
-            var day = value.Day;
-            if (year < 10)
-            {
-                writer.WriteRawUnsafe((byte)'0');
-                writer.WriteRawUnsafe((byte)'0');
-                writer.WriteRawUnsafe((byte)'0');
-            }
-            else if (year < 100)
-            {
-                writer.WriteRawUnsafe((byte)'0');
-                writer.WriteRawUnsafe((byte)'0');
-            }
-            else if (year < 1000)
-            {
-                writer.WriteRawUnsafe((byte)'0');
-            }
-            writer.WriteInt32(year);
-            writer.WriteRawUnsafe((byte)'-');
-
-            if (month < 10)
-            {
-                writer.WriteRawUnsafe((byte)'0');
-            }
-            writer.WriteInt32(month);
-            writer.WriteRawUnsafe((byte)'-');
-
-            if (day < 10)
-            {
-                writer.WriteRawUnsafe((byte)'0');
-            }
-            writer.WriteInt32(day);
+            WriteDate(ref writer, value, formatterResolver);
             writer.WriteRawUnsafe((byte)'\"');
         }
 
         public LocalDate Deserialize(ref JsonReader reader, IJsonFormatterResolver formatterResolver)
         {
             var str = reader.ReadStringSegmentUnsafe();
-            var array = str.Array;
             var i = str.Offset;
-            var len = str.Count;
+            var date = ReadDate(str, formatterResolver, ref i);
 
-            switch (len)
-            {
-                // YYYY
-                case 4:
-                {
-                    var y = (array[i++] - (byte)'0') * 1000 + (array[i++] - (byte)'0') * 100 + (array[i++] - (byte)'0') * 10 + (array[i++] - (byte)'0');
-                    return new LocalDate(y, 1, 1);
-                }
-                // YYYY-MM
-                case 7:
-                {
-                    var y = (array[i++] - (byte)'0') * 1000 + (array[i++] - (byte)'0') * 100 + (array[i++] - (byte)'0') * 10 + (array[i++] - (byte)'0');
-                    if (array[i++] != (byte)'-') Exceptions.ThrowInvalidDateTimeFormat(str);
-                    var m = (array[i++] - (byte)'0') * 10 + (array[i++] - (byte)'0');
-                    return new LocalDate(y, m, 1);
-                }
-                // YYYY-MM-DD
-                case 10:
-                {
-                    var y = (array[i++] - (byte)'0') * 1000 + (array[i++] - (byte)'0') * 100 + (array[i++] - (byte)'0') * 10 + (array[i++] - (byte)'0');
-                    if (array[i++] != (byte)'-') Exceptions.ThrowInvalidDateTimeFormat(str);
-                    var m = (array[i++] - (byte)'0') * 10 + (array[i++] - (byte)'0');
-                    if (array[i++] != (byte)'-') Exceptions.ThrowInvalidDateTimeFormat(str);
-                    var d = (array[i++] - (byte)'0') * 10 + (array[i++] - (byte)'0');
-                    return new LocalDate(y, m, d);
-                }
-                default:
-                    Exceptions.ThrowInvalidDateTimeFormat(str);
-                    break;
-            }
-
-            return default(LocalDate);
+            return date;
         }
     }
 }
